requirements.txt
discord.py==2.3.2
google-generativeai==0.8.5
python-dotenv==1.0.0
requests==2.32.5